Revert "Stop using String names for LinkLayer/DLT types"
